Use BitSliceIterator
