+++ arrow-data/src/equal/primitive.rs
@@ -15,12 +15,12 @@
 // specific language governing permissions and limitations
 // under the License.
 
+use crate::bit_iterator::BitSliceIterator;
 use crate::contains_nulls;
 use arrow_buffer::bit_util::get_bit;
 use std::mem::size_of;
 
 use crate::data::ArrayData;
-use crate::slices_iterator::SlicesIterator;
 
 use super::utils::equal_len;
 
@@ -73,12 +73,12 @@ pub(super) fn primitive_equal<T>(
                         )
             })
         } else {
-            let lhs_slices_iter = SlicesIterator::new_from_buffer(
+            let lhs_slices_iter = BitSliceIterator::new(
                 lhs.null_buffer().as_ref().unwrap(),
                 lhs_start + lhs.offset(),
                 len,
             );
-            let rhs_slices_iter = SlicesIterator::new_from_buffer(
+            let rhs_slices_iter = BitSliceIterator::new(
                 rhs.null_buffer().as_ref().unwrap(),
                 rhs_start + rhs.offset(),
                 len,
